requirements.txt
langchain
langchain-google-genai
langchain-community
langchain-core
python-dotenv
SQLAlchemy
PyMySQL  # MySQL driver for SQLAlchemy
python-dateutil>=2.8.2  # For date handling
pydantic>=2.0.0  # For data validation